backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
stripe==7.8.0
python-dotenv==1.0.0
fpdf2==2.7.6
google-generativeai==0.3.1
pyjwt==2.8.0
requests==2.31.0
